API/Web/requirements.txt
Flask
flask-sqlalchemy
PyMySQL
Flask-Limiter
python-dotenv
gunicorn
tensorflow==2.4.1
keras==2.4.0
transformers==4.2.2
PySastrawi==1.2.0
numpy
pandas
